backend/requirements.txt
fastapi[standard]==0.113.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pyjwt==2.10.1
bcrypt==4.0.1
passlib[bcrypt]==1.7.4